Fix where the userCount is in the message (#68)
* Fix where the userCount is in the message

The message was changed, but the code which parses the message later
wasn't updated. Now that location is in a const which should make it
more searchable in the future.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Use regex to extract usercount from violation

Previously we split the string on spaces, then relied on the number
being at a specific index. Since the role name can have spaces, that
technique is not correct.

I believe `MustCompile` is ok here - we control the string coming in,
we can guarantee it does not panic.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Refactor extracting user count into function

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Remove some dead code

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Tiny updates to trigger prow rebuild

Hopefully, this should invalidate a possible image layer cache that
prow is using that (might) be interfering with its builds.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>