Account for OpenShift groups when counting the users with a role (#66)
When examining which users have a role, group membership must also be
considered. Group membership can only be determined if the group is an
OpenShift Group kind. Otherwise, it is up to the external authentication
system configured with the Kubernetes cluster to provide this value when
the user logs in. In the latter case, there isn't much that can be done.

Resolves https://github.com/open-cluster-management/backlog/issues/14166

Signed-off-by: mprahl <mprahl@users.noreply.github.com>