Add support for modifying what cluster role bindings are ignored (#73)
Before this commit, any cluster role binding that started with "system"
was ignored. The intention was to ignore any default cluster role
bindings generated by Kubernetes/OpenShift. This was too liberal since
it also ignored cluster role bindings named things like
"system-administrators". The default is thus changed to "system:".

Since there is nothing restricting users of creating cluster role bindings
that start with "system:", a new field on the IamPolicy kind was added
called "spec.ignoreClusterRoleBindings". This is an array of regex
values that can be used to ignore cluster role bindings that the
user desires. This means they can have a stricter regex or disable it
entirely with a regex such as `.^`.

Resolves: https://github.com/open-cluster-management/backlog/issues/14874

Signed-off-by: mprahl <mprahl@users.noreply.github.com>