Fix operator-sdk manifest generation
The `CompliancyDetails` type was an invalid type for manifest generation
and needed a custom type for its map value.

Additionally, for the operator-sdk command to run, the `Dockerfile`
had to be moved to the `build` directory or else it exits with an error
that it's not in the project root.

Additionally, the code and manifest generation was rerun. This removed
the `minLength: 1` option on `ignoreClusterRoleBindings` but this will
be handled in a future commit.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>