Add field validation to the IamPolicy CRD (#74)
* Fix a typo in the IamPolicy example

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Fix operator-sdk manifest generation

The `CompliancyDetails` type was an invalid type for manifest generation
and needed a custom type for its map value.

Additionally, for the operator-sdk command to run, the `Dockerfile`
had to be moved to the `build` directory or else it exits with an error
that it's not in the project root.

Additionally, the code and manifest generation was rerun. This removed
the `minLength: 1` option on `ignoreClusterRoleBindings` but this will
be handled in a future commit.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Fix the make targets for operator-sdk code generation

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Add field validation to the IamPolicy CRD

This specifically targeted the user-facing fields.

Most field validation occurs client and server side except for unknown
fields. That is handled only on the client side validation. This is
enabled by default with `kubectl`, however, `oc` requires `--validate=true`
to be set. When an unknown field is provided with client side validation
disabled, the server will ignore it (default for v1 CRDs) and not store
the unknown field in etcd.

Note that `additionalProperties: false` cannot be set on objects which have
the `properties` validation. This is a restriction by Kubernetes. The
workaround is setting `spec.preserveUnknownFields: false` (default for
v1 CRDs) and having client side validation enabled.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Add a symlink for the Dockerfile in the root

This is required until it is adjusted in the Prow configuration:
https://github.com/openshift/release/blob/master/ci-operator/config/open-cluster-management/iam-policy-controller/open-cluster-management-iam-policy-controller-release-2.4.yaml

Signed-off-by: mprahl <mprahl@users.noreply.github.com>