Upgrade to the latest operator-sdk (#78)
* Remove stale code about namespaceSelector

Since namespaceSelector does not affect the policy controller's
decisions, the helper code was removed and the field in the CRD was
marked as obsolete.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Upgrade to the latest operator-sdk

operator-sdk v1.x now wraps Kubebuilder for its scaffolding and code
generation. This changes the path of `./cmd/manager/main.go` to
`.main.go`, `./pkg/controller` to `./controllers`, and `pkg/apis` to
`api`.

Manifest generation is also handled by Kubebuilder. This is handled by
the `generate-operator-yaml` Make target. This means that any RBAC
changes must be done using kubebuilder markers and not manual edits to
the manifests in `./deploy`. The RBAC Kubebuilder markers have been
refined to follow the least privilege principal.

The largest changes stem from updating the underlying Kubernetes Go
modules as some APIs changed or were removed. This also leads to a new
way of performing leader election via leases. For old Kubernetes
versions such as OCP 3.11, the lease API is not available so a flag
called `--legacy-leader-election` will go back to the old way of
performing leader election.

It's also important to note that operator-sdk is no longer a Go module
dependency, so some functionality was copied over, mostly unmodified.

Lastly, the Policy scheme is now imported from the propagator rather
than redefined in this controller.

Relates:
https://github.com/open-cluster-management/backlog/issues/15552

Signed-off-by: mprahl <mprahl@users.noreply.github.com>