Use utility package for logging (#102)
* Remove unused watch-ns flag

The value in this flag was being overwritten by common.GetWatchNamespace
when it was being provided.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Configure logging with utility package

This will make the logs look more consistent, and allow for better run-
time configuration of both logs this controller creates, and the logs
that libraries might create.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Configure logging with utility package

This will make the logs look more consistent, and allow for better run-
time configuration of both logs this controller creates, and the logs
that libraries might create.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>